requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
urllib3==2.2.3
pydantic==2.10.1
langchain==0.3.9
langchain-openai==0.2.10
python-dotenv==1.0.0 
pymupdf==1.24.14 